+++ editor/command/type/DeleteEntityCmd.cpp
@@ -48,7 +48,7 @@ editor::DeleteEntityCmd::DeleteEntityCmd(Project* project, uint32_t sceneId, con
 }
 
 void editor::DeleteEntityCmd::destroyEntity(EntityRegistry* registry, Entity entity, std::vector<Entity>& entities, Project* project, uint32_t sceneId){
-    int structuralFlags = UpdateFlags_None;
+    uint64_t structuralFlags = UpdateFlags_None;
     if (registry->isEntityCreated(entity)){ // locked child are deleted by systems when their parent is deleted
         if (registry->findComponent<LightComponent>(entity)){
             structuralFlags |= Catalog::getComponentStructuralUpdateFlags(ComponentType::LightComponent);
